@@ -19,7 +19,7 @@
         "irc": "irc://irc.freenode.org/sf-grenoble"
     },
     "require": {
-        "php": ">=7.4",
+        "php": ">=8.0",
         "symfony/config": "^5.0|^6.0",
         "symfony/dependency-injection": "^5.0|^6.0",
         "symfony/http-kernel": "^5.0|^6.0",
