@@ -2,19 +2,18 @@
 
 namespace BOMO\IcalBundle\Model;
 
+use DateTimeInterface;
+use Kigkonsult\Icalcreator\Pc;
 use Kigkonsult\Icalcreator\Vevent;
 
 class Event
 {
-    /**
-     * vEvent object
-     */
-    private $event;
+    private vEvent $event;
 
     /**
      * To ignore hours
      */
-    private $isAllDayEvent;
+    private bool $isAllDayEvent;
 
     public function __construct($object = null)
     {
@@ -25,7 +24,7 @@ class Event
         }
     }
 
-    public function setStartDate(\Datetime $date)
+    public function setStartDate(null|string|DateTimeInterface|Pc $date): static
     {
         $params = array();
         if (true === $this->isAllDayEvent) {
@@ -37,7 +36,7 @@ class Event
         return $this;
     }
 
-    public function setEndDate(\Datetime $date)
+    public function setEndDate(null|string|DateTimeInterface|Pc $date): static
     {
         $params = array();
         if (true === $this->isAllDayEvent) {
@@ -49,96 +48,96 @@ class Event
         return $this;
     }
 
-    public function setIsAllDayEvent($bool = true)
+    public function setIsAllDayEvent(bool $bool = true): static
     {
         $this->isAllDayEvent = $bool;
 
         return $this;
     }
 
-    public function setName($name)
+    public function setName(null|string|Pc $name): static
     {
         $this->event->setSummary($name);
 
         return $this;
     }
 
-    public function setLocation($loc)
+    public function setLocation(null|string|Pc $loc): static
     {
         $this->event->setLocation($loc);
 
         return $this;
     }
 
-    public function setDescription($desc)
+    public function setDescription(null|string|Pc $desc): static
     {
         $this->event->setDescription($desc);
 
         return $this;
     }
 
-    public function setComment($comment)
+    public function setComment(null|string|Pc $comment): static
     {
         $this->event->setComment($comment);
 
         return $this;
     }
 
-    public function setAttendee($attendee)
+    public function setAttendee(null|string|Pc $attendee): static
     {
         $this->event->setAttendee($attendee);
 
         return $this;
     }
 
-    public function setOrganizer($organizer)
+    public function setOrganizer(null|string|Pc $organizer): static
     {
         $this->event->setOrganizer($organizer);
 
         return $this;
     }
 
-    public function setStatus($status)
+    public function setStatus(null|string|Pc $status): static
     {
         $this->event->setStatus($status);
 
         return $this;
     }
 
-    public function setTransparent($name)
+    public function setTransparent(null|string|Pc $name): static
     {
         $this->event->setTransp($name);
 
         return $this;
     }
 
-    public function setPriority($value)
+    public function setPriority(null|int|string|Pc $value): static
     {
         $this->event->setPriority($value);
 
         return $this;
     }
 
-    public function setSequence($value)
+    public function setSequence(null|int|string|Pc $value): static
     {
         $this->event->setSequence($value);
 
         return $this;
     }
 
-    public function setUrl($url)
+    public function setUrl(null|string|Pc $url): static
     {
         $this->event->setUrl($url);
 
         return $this;
     }
 
-    public function newAlarm()
+    public function newAlarm(): Alarm
     {
-        return new Alarm($this->event->newValarm(), $this->event);
+        return new Alarm($this->event->newValarm());
     }
 
-    public function attachAlarm(Alarm $alarm)
+    public function attachAlarm(Alarm $alarm): static
     {
         $this->event->setComponent($alarm->getAlarm());
 
@@ -152,12 +151,12 @@ class Event
         }
     }
 
-    public function getEvent()
+    public function getEvent(): Vevent
     {
         return $this->event;
     }
 
-    private function datetimeToArray(\Datetime $datetime)
+    private function datetimeToArray(\Datetime $datetime): array
     {
         $str = $datetime->format('Y-m-d H:i:s');
 
